requirements.txt
Django==2.2
django-filter==2.4.0
django-rest-framework==0.1.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
djoser==2.1.0
pkg-resources==0.0.0
pytz==2021.1
sqlparse==0.4.1